+++ src/steps/compute/index.ts
@@ -337,7 +337,7 @@ export const computeSteps: IntegrationStep<IntegrationConfig>[] = [
       },
     ],
     relationships: [],
-    executionHandler: withErrorHandling(fetchComputeNetworks),
+    executionHandler: fetchComputeNetworks,
   },
   {
     id: STEP_COMPUTE_FIREWALLS,
@@ -369,7 +369,7 @@ export const computeSteps: IntegrationStep<IntegrationConfig>[] = [
         targetType: ENTITY_TYPE_COMPUTE_FIREWALL,
       },
     ],
-    executionHandler: withErrorHandling(fetchComputeFirewalls),
+    executionHandler: fetchComputeFirewalls,
     dependsOn: [STEP_COMPUTE_NETWORKS],
   },
   {
@@ -390,7 +390,7 @@ export const computeSteps: IntegrationStep<IntegrationConfig>[] = [
         targetType: ENTITY_TYPE_COMPUTE_SUBNETWORK,
       },
     ],
-    executionHandler: withErrorHandling(fetchComputeSubnetworks),
+    executionHandler: fetchComputeSubnetworks,
     dependsOn: [STEP_COMPUTE_NETWORKS],
   },
   {
@@ -404,7 +404,7 @@ export const computeSteps: IntegrationStep<IntegrationConfig>[] = [
       },
     ],
     relationships: [],
-    executionHandler: withErrorHandling(fetchComputeDisks),
+    executionHandler: fetchComputeDisks,
   },
   {
     id: STEP_COMPUTE_INSTANCES,
@@ -442,6 +442,6 @@ export const computeSteps: IntegrationStep<IntegrationConfig>[] = [
       STEP_COMPUTE_NETWORKS,
       STEP_COMPUTE_SUBNETWORKS,
     ],
-    executionHandler: withErrorHandling(fetchComputeInstances),
+    executionHandler: fetchComputeInstances,
   },
 ];